+++ CHANGELOG.md
@@ -8,14 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
 - Add `SingleNodeAudioTransformDevice` to make simple audio transforms easier to write.
+- Add End-to-end Integration test for Video Test App
 
 ### Changed
 
 ### Removed
 
 ### Fixed
+
 
 ## [2.5.0] - 2021-02-16